refactor(noun): establishGender fn
